cloud-funcs/http-api/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore==2.3.3
google-cloud-pubsub==2.8.0